Extract FakeProvider for global use among packages
The FakeProvider class is useful in tests because it allows you to pass
a provider that doesn't make HTTP requests to a library such as
`eth-query`.

Previously, this class was only available in NetworkController tests.
With this commit, you can use it anywhere.
